+++ app/ids-loading-indicator/side-by-side.js
@@ -1,0 +1,11 @@
+/* global $ */
+
+import IdsLoadingIndicator from '../../src/ids-loading-indicator/ids-loading-indicator';
+
+// Initialize the 4.x
+$('body').initialize();
+$(() => {
+  $('#submit').click((e) => { // eslint-disable-line no-unused-vars
+    $('#busy-form').trigger('start.busyindicator');
+  });
+});