+++ doc/CHANGELOG.md
@@ -17,6 +17,7 @@
   - If using events, events are now plain JS events.
   - Can now be imported as a single JS file and used with encapsulated styles
 - `[Colorpicker]` The color picker has been ported from 4.x and converted to a custom element `<ids-color-picker></ids-color-picker>`
+- `[IdsContainer]` Adding a new setting to provide a custom padding property. <ids-container padding="8"></ids-container>
 - `[Counts]` The counts component has been changed to a web component and renamed to ids-counts.
   - Text is now contained in an ids-text element `<ids-text></ids-text>`
   - Can now be imported as a single JS file and used with encapsulated styles